+++ src/viewpanel/components/SeoDetails.tsx
@@ -1,21 +1,41 @@
 import * as React from 'react';
+import { VsTable, VsTableBody, VsTableHeader, VsTableHeaderCell, VsTableRow, VsTableCell } from './VscodeComponents';
 
 export interface ISeoDetailsProps {
   allowedLength: number;
   title: string;
-  value: string;
+  value: number;
+  valueTitle: string;
+  noValidation?: boolean;
 }
 
 export const SeoDetails: React.FunctionComponent<ISeoDetailsProps> = (props: React.PropsWithChildren<ISeoDetailsProps>) => {
-  const { allowedLength, title, value } = props;
+  const { allowedLength, title, value, valueTitle, noValidation } = props;
+
+  const validate = () => {
+    if (noValidation) {
+      return "";
+    }
+
+    return value <= allowedLength ? "valid" : "not-valid"
+  };
 
   return (
-    <div className={`seo__status__details ${value.length <= allowedLength ? "valid" : "not-valid"}`}>
-      <h4><strong>{title}</strong></h4>
-      <ul>
-        <li><b>Length</b>: {value.length}</li>
-        <li><b>Recommended length</b>: {allowedLength}</li>
-      </ul>
+    <div className={`seo__status__details ${validate()}`}>
+      <h4>{title}</h4>
+
+      <VsTable bordered>
+        <VsTableHeader slot="header">
+          <VsTableHeaderCell className={validate()}>{valueTitle}</VsTableHeaderCell>
+          <VsTableHeaderCell>Recommended</VsTableHeaderCell>
+        </VsTableHeader>
+        <VsTableBody slot="body">
+          <VsTableRow>
+            <VsTableCell className={validate()}>{value}</VsTableCell>
+            <VsTableCell>{allowedLength}</VsTableCell>
+          </VsTableRow>
+        </VsTableBody>
+      </VsTable>
     </div>
   );
 };