+++ src/viewpanel/components/VscodeComponents.ts
@@ -1,3 +1,9 @@
 import {wrapWc} from 'wc-react';
 
-export const VscodeIcon = wrapWc(`vscode-icon`);
+export const VsTable = wrapWc(`vscode-table`);
+export const VsTableHeader = wrapWc(`vscode-table-header`);
+export const VsTableHeaderCell = wrapWc(`vscode-table-header-cell`);
+export const VsTableBody = wrapWc(`vscode-table-body`);
+export const VsTableRow = wrapWc(`vscode-table-row`);
+export const VsTableCell = wrapWc(`vscode-table-cell`);
+export const VsCollapsible = wrapWc(`vscode-collapsible`);